requirements.txt
pytest
pytest_bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
allure-pytest
lxml-4.6.4-cp39-cp39-win_amd64.whl
